src/api/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
PyJWT==2.8.0
pytz==2024.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
urllib3==2.2.1
